# openapi.yml
openapi: "3.0.0"
info:
  version: v0
  title: youyuan core Service
  license:
    name: MIT
servers:
  - url: https://api.youyuan.36node.com{basePath}
    description: The production API server
    variables:
      basePath:
        default: /core/v0
paths:
  /parks:
    post:
      summary: Create a park
      operationId: createPark
      tags:
        - park
      requestBody:
        description: Park to add
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParkCreateDoc"
      responses:
        "201":
          description: The Park created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Park"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"

    get:
      summary: List all parks
      operationId: listParks
      tags:
        - park
      parameters:
        - name: _limit
          in: query
          description: How many items to return at one time (max 100)
          required: false
          schema:
            type: integer
            format: int32
            default: 10
            maximum: 1000
        - name: _offset
          in: query
          description: How many items to escape
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: _sort
          in: query
          description: Fields to sort
          required: false
          schema:
            type: string
        - name: _select
          in: query
          description: Fields to select
          required: false
          schema:
            type: array
            items:
              type: string
        - name: name_like
          in: query
          description: name like of park
          required: false
          schema:
            type: string
      responses:
        "200":
          description: A paged array of parks
          headers:
            X-Total-Count:
              description: Total count of all warnings match the query
              required: true
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Park"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /parks/{parkId}:
    get:
      summary: Find park by id
      operationId: getPark
      tags:
        - park
      parameters:
        - name: parkId
          in: path
          required: true
          description: The id of the park to retrieve
          schema:
            type: string
      responses:
        "200":
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Park"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    put:
      summary: Update park
      operationId: updatePark
      tags:
        - park
      parameters:
        - name: parkId
          in: path
          required: true
          description: The id of park
          schema:
            type: string
      requestBody:
        description: Park to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParkDoc"
      responses:
        "200":
          description: The park
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Park"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      description: deletes a single park based on the ID supplied
      operationId: deletePark
      tags:
        - park
      parameters:
        - name: parkId
          in: path
          description: ID of park to delete
          required: true
          schema:
            type: string
      responses:
        "204":
          description: park deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /maintains:
    post:
      summary: Create a maintain order
      operationId: createMaintain
      tags:
        - maintain
      requestBody:
        description: maintain order to add
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaintainDoc"
      responses:
        "201":
          description: The maintain order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Maintain"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"

    get:
      summary: List all maintain orders
      operationId: listMaintains
      tags:
        - maintain
      parameters:
        - name: _limit
          in: query
          description: How many items to return at one time (max 100)
          required: false
          schema:
            type: integer
            format: int32
            default: 10
            maximum: 1000
        - name: _offset
          in: query
          description: How many items to escape
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: _sort
          in: query
          description: Fields to sort
          required: false
          schema:
            type: string
        - name: _select
          in: query
          description: Fields to select
          required: false
          schema:
            type: array
            items:
              type: string
        - name: id_like
          in: query
          description: repair order id, regix
          required: false
          schema:
            type: string
            minLength: 1
        - name: status
          in: query
          description: maintain order status
          required: false
          schema:
            type: string
            enum: ["CREATING", "WAITING", "MAINTAINING", "REPARING", "MAINTAINED", "PENDING", "PASSED"]
        - name: createAt_gt
          in: query
          description: greater than createAt
          required: false
          schema:
            type: string
            format: date-time
        - name: createAt_lt
          in: query
          description: less than createAt
          required: false
          schema:
            type: string
            format: date-time
        - name: terminal_like
          in: query
          description: terminal, regix
          required: false
          schema:
            type: string
            minLength: 1
        - name: plate_like
          in: query
          description: plate, regix
          required: false
          schema:
            type: string
            minLength: 1
        - name: crm_like
          in: query
          description: crm, regix
          required: false
          schema:
            type: string
            minLength: 1
        - name: vin_like
          in: query
          description: vin, regix
          required: false
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: A paged array of maintain orders
          headers:
            X-Total-Count:
              description: Total count of all warnings match the query
              required: true
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Maintain"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /maintains/{maintainId}:
    get:
      summary: Find maintain order by id
      operationId: getMaintain
      tags:
        - maintain
      parameters:
        - name: maintainId
          in: path
          required: true
          description: The id of the maintain order to retrieve
          schema:
            type: string
      responses:
        "200":
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Maintain"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    put:
      summary: update maintain order
      operationId: updateMaintain
      tags:
        - maintain
      parameters:
        - name: maintainId
          in: path
          required: true
          description: The id of maintain order
          schema:
            type: string
      requestBody:
        description: maintain order to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaintainDoc"
      responses:
        "200":
          description: The maintain order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Maintain"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      description: deletes a single maintain order based on the ID supplied
      operationId: deleteMaintain
      tags:
        - maintain
      parameters:
        - name: maintainId
          in: path
          description: ID of maintain order to delete
          required: true
          schema:
            type: string
      responses:
        "204":
          description: maintain order deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /maintains/{maintainId}/appearanceRecord:
    put:
      summary: update maintain order appearance record
      operationId: updateAppearanceRecord
      tags:
        - maintain
      parameters:
        - name: maintainId
          in: path
          required: true
          description: The id of maintain order
          schema:
            type: string
      requestBody:
        description: maintain order appearance record to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaintainAppearanceRecord"
      responses:
        "200":
          description: The maintain order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintainAppearanceRecord"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /maintains/{maintainId}/softwareRecord:
    put:
      summary: update maintain order software record
      operationId: updateSoftwareRecord
      tags:
        - maintain
      parameters:
        - name: maintainId
          in: path
          required: true
          description: The id of maintain order
          schema:
            type: string
      requestBody:
        description: maintain order software record to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaintainSoftwareRecord"
      responses:
        "200":
          description: The maintain order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintainSoftwareRecord"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
security:
  - bearerAuth: []
components:
  responses:
...
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Err"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Err"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Err"
    InternalError:
      description: unexpected error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Err"
  schemas:
    ParkDoc:
      description: 车场信息
      type: object
      properties:
        name:
          type: string
          description: 车场名
        address:
          type: string
          description: 车场地址
        appearance:
          type: string
          description: 外观检测默认组ns
        software:
          type: string
          description: 软件诊断默认组ns
    ParkCreateDoc:
      allOf:
        - $ref: "#/components/schemas/ParkDoc"
        - type: object
          required:
            - name
            - address
            - appearance
            - software
          properties:
            name:
              type: string
            address:
              type: string
            appearance:
              type: string
            software:
              type: string
    Park:
      allOf:
        - $ref: "#/components/schemas/ParkDoc"
        - $ref: "#/components/schemas/MongoDefault"
...
    MaintainAppearanceRecord:
      type: object
      description: 维保单外观诊断记录
      required:
        - appearanceSignAddress
        - appearanceArrivalAt
        - appearanceWorkers
        - meterAlarm
        - boxDamage
        - electricLeakage
        - valveLoose
        - coverAbnormal
        - headLoose
        - boltLoose
        - msdAbnormal
        - hvWireWear
        - lvWireWear
        - hvConnectorFasten
        - lvConnectorFasten
        - frontPic
        - nameplatePic
        - meterPic
        - batteryPic
      properties:
        appearanceSignAddress:
          type: string
        appearanceArrivalAt:
          type: string
          format: date-time
        appearanceRouteMap:
          type: string
        appearanceWorkers:
          type: array
          items:
            type: string
        meterAlarm:
          type: boolean
        meterAlarmResult:
          type: boolean
        meterAlarmDesc:
          type: string
        boxDamage:
          type: boolean
        boxDamageResult:
          type: boolean
        boxDamageDesc:
          type: string
        electricLeakage:
          type: boolean
        electricLeakageResult:
          type: boolean
        electricLeakageDesc:
          type: string
        valveLoose:
          type: boolean
        valveLooseResult:
          type: boolean
        valveLooseDesc:
          type: string
        coverAbnormal:
          type: boolean
        coverAbnormalResult:
          type: boolean
        coverAbnormalDesc:
          type: string
        headLoose:
          type: boolean
        headLooseResult:
          type: boolean
        headLooseDesc:
          type: string
        boltLoose:
          type: boolean
        boltLooseResult:
          type: boolean
        boltLooseDesc:
          type: string
        msdAbnormal:
          type: boolean
        msdAbnormalResult:
          type: boolean
        msdAbnormalDesc:
          type: string
        hvWireWear:
          type: boolean
        hvWireWearResult:
          type: boolean
        hvWireWearDesc:
          type: string
        lvWireWear:
          type: boolean
        lvWireWearResult:
          type: boolean
        lvWireWearDesc:
          type: string
        hvConnectorFasten:
          type: boolean
        hvConnectorFastenResult:
          type: boolean
        hvConnectorFastenDesc:
          type: string
        lvConnectorFasten:
          type: boolean
        lvConnectorFastenResult:
          type: boolean
        lvConnectorFastenDesc:
          type: string
        frontPic:
          type: array
          items:
            type: string
        nameplatePic:
          type: array
          items:
            type: string
        meterPic:
          type: array
          items:
            type: string
        batteryPic:
          type: array
          items:
            type: string
        status:
          type: string
          enum: ["CREATING", "WAITING", "MAINTAINING", "REPARING", "PENDING", "PASSED"]
    MaintainSoftwareRecord:
      type: object
      description: 维保单软件诊断记录
      required:
        - softwareSignAddress
        - softwareArrivalAt
        - softwareWorkers
        - softwareStartPic
        - softwareEndPic
        - soh
        - current
        - soc
        - roc
        - addVoltage
        - totalVoltage
        - maxVoltage
        - minVoltage
        - maxTemperature
        - minTemperature
        - diffTemperature
        - batteryBalance
        - diffVoltage
        - chargePositive
        - chargeNegative
        - nochargePositive
        - nochargeNegative
      properties:
        softwareSignAddress:
          type: string
        softwareArrivalAt:
          type: string
          format: date-time
        softwareRouteMap:
          type: string
        softwareWorkers:
          type: array
          items:
            type: string
        softwareStartPic:
          type: array
          items:
            type: string
        softwareEndPic:
          type: array
          items:
            type: string
        soh:
          type: string
        sohResult:
          type: string
        sohDesc:
          type: string
        current:
          type: string
        currentResult:
          type: string
        currentDesc:
          type: string
        soc:
          type: string
        socResult:
          type: string
        roc:
          type: string
        rocResult:
          type: string
        socDesc:
          type: string
        addVoltage:
          type: string
        addVoltageResult:
          type: string
        totalVoltage:
          type: string
        totalVoltageResult:
          type: string
        addVoltageDesc:
          type: string
        maxVoltage:
          type: string
        maxVoltageResult:
          type: string
        maxVoltageDesc:
          type: string
        minVoltage:
          type: string
        minVoltageResult:
          type: string
        minVoltageDesc:
          type: string
        maxTemperature:
          type: string
        maxTemperatureResult:
          type: string
        maxTemperatureDesc:
          type: string
        minTemperature:
          type: string
        minTemperatureResult:
          type: string
        minTemperatureDesc:
          type: string
        diffTemperature:
          type: string
        diffTemperatureResult:
          type: string
        diffTemperatureDesc:
          type: string
        batteryBalance:
          type: string
          enum: ["LOW", "MIDDLE", "HIGH"]
        diffVoltage:
          type: string
        diffVoltageResult:
          type: string
        diffVoltageDesc:
          type: string
        chargePositive:
          type: string
        chargePositiveResult:
          type: string
        chargeNegative:
          type: string
        chargeNegativeResult:
          type: string
        chargePositiveDesc:
          type: string
        nochargePositive:
          type: string
        nochargePositiveResult:
          type: string
        nochargeNegative:
          type: string
        nochargeNegativeResult:
          type: string
        nochargePositiveDesc:
          type: string
        status:
          type: string
          enum: ["CREATING", "WAITING", "MAINTAINING", "REPARING", "PENDING", "PASSED"]
    MaintainDoc:
      type: object
      properties:
        ticket:
          type: string
          description: 关联的工单 id
        status:
          type: string
          enum: ["CREATING", "WAITING", "MAINTAINING", "REPARING", "PENDING", "PASSED"]
          description: 维保单工单状态
        rejected:
          type: boolean
          description: 工单曾经被驳回
        crm:
          type: string
          description: CRM 订单号
        vehicle:
          type: string
          description: 关联车辆 id
        terminal:
          type: string
          description: 关联车辆 终端用户
        plate:
          type: string
          description: 关联车辆 车牌号
        vin:
          type: string
          description: 关联车辆 vin
        brands:
          type: string
          description: 关联车辆 整车品牌
        appearance:
          type: string
          description: 外观检查组
        software:
          type: string
          description: 软件诊断组
        closeAt:
          type: string
          format: date-time
          description: 完工时间
        appearanceRecord:
          $ref: "#/components/schemas/MaintainAppearanceRecord"
        softwareRecord:
          $ref: "#/components/schemas/MaintainSoftwareRecord"
        repair:
          type: string
          description: 关联维修单 id
    Maintain:
      allOf:
        - $ref: "#/components/schemas/MaintainDoc"
        - $ref: "#/components/schemas/MongoDefault"
...
    MongoDefault:
      type: object
      required:
        - id
      properties:
        id:
          type: string
        updateAt:
          type: string
          format: date-time
        updateBy:
          type: string
        createAt:
          type: string
          format: date-time
        createBy:
          type: string
    Err:
      type: object
      required:
        - name
        - message
      properties:
        code:
          type: string
        type:
          type: string
        message:
          type: boolean
        name:
          type: string
        details:
          type: array
          items:
            type: object
            properties:
              keyword:
                type: string
              message:
                type: string
              path:
                type: string
              value:
                type: string
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT